фикс миграции

+++ internal/storage/migrations/000003_create_withdrawals_table.up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS withdrawals (
     user_login TEXT,
     order_id TEXT,
-    withdrawal_sum INTEGER,
+    withdrawal_sum DECIMAL,
     processed_at TIMESTAMP
 )
